+++ safeDelivery/src/com/safeDelivery/service/VilleService.java
@@ -3,6 +3,6 @@ package com.safeDelivery.service;
 import com.safeDelivery.model.Ville;
 
 public interface VilleService {
-	public int existByName(String nom);
-	public int saveVille(Ville ville);
+	public long existByName(String nom);
+	public long saveVille(Ville ville);
 }
