+++ safeDelivery/src/com/safeDelivery/service/ZoneService.java
@@ -1,9 +1,8 @@
 package com.safeDelivery.service;
 
-import com.safeDelivery.model.Ville;
 import com.safeDelivery.model.Zone;
 
 public interface ZoneService {
-	public int existByName(String nom);
-	public int saveZone(Zone zone);
+	public long existByName(String nom);
+	public long saveZone(Zone zone);
 }
